Check added to ensure that .horn is the root of the package tree.

+++ trunk/src/Horn.Core/SourceControl/SvnSourceControl.cs
@@ -45,7 +45,8 @@ namespace Horn.Core.SCM
 
         protected override void Initialise(IPackageTree packageTree)
         {
-
+            if(!packageTree.Root.Name.StartsWith(PackageTree.RootPackageTreeName))
+                throw new InvalidOperationException("The root of the package tree is not named horn");
 
             if (packageTree.WorkingDirectory.Exists)
                 packageTree.WorkingDirectory.Delete(true);
